changed compressed ptr to 32 bits and slab indx size to 15 bits

+++ cachelib/allocator/datastruct/tests/SListTest.cpp
@@ -35,11 +35,11 @@ class CACHELIB_PACKED_ATTR SListNode {
     // default construct to nullptr.
     CompressedPtr() = default;
 
-    explicit CompressedPtr(int64_t ptr) : ptr_(ptr) {}
+    explicit CompressedPtr(int32_t ptr) : ptr_(ptr) {}
 
-    int64_t saveState() const noexcept { return ptr_; }
+    int32_t saveState() const noexcept { return ptr_; }
 
-    int64_t ptr_{};
+    int32_t ptr_{};
   };
 
   struct PtrCompressor {
